+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## vNext
 
 * Add `pwt stats verbose` to log the server output when running the stats
+* Downgrade ts-jest to 20.0.4 due to performance regression
+  (https://github.com/kulshekhar/ts-jest/issues/259)
 
 ## v0.6.0 (2017-07-27)