+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `pwt stats verbose` to log the server output when running the stats
 * Downgrade ts-jest to 20.0.4 due to performance regression
   (https://github.com/kulshekhar/ts-jest/issues/259)
+* Fix webpack performance regression by reducing source map resolution and stats output in dev
 
 ## v0.6.0 (2017-07-27)